GrpcClientRequest can sometimes reset the HTTP/2 stream when it receives the gRPC response.
GrpcClientRequest does not properly mark the request as ended when it has sent the request, leading to resetting the stream when it receives the response.

GrpcClientRequest has been modified to mark the request as sent in all situations. In addition the implementation of reset has been improved to only reset the stream when the stream is not closed.

fixes #40