Avoid false-positive leak detection
See #98

In most cases, the encoded message buffer is sent to the wire directly, and thus it is not GC-ed before it's actually written and release at lower level by Netty.

However, with gRPC-Web in text mode, the buffer content is base64 encoded and the initial composite ByteBuf is never released, even if it's safe not to do so because it's not pooled nor direct.

To avoid the false-positive, the composite byte buff aggregation can be replaced with Vert.x Buffer manipulation. In this part of the code we're only dealing with un-pooled heap buffers anyway.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>