Avoid masking invalid grpc status errors in GrpcClientResponse#last().
Motivation

The current handling of some futures of GrpcClientResponse will map the end future to a failure when the status is not OK with a mere message error.

We can instead more precisely handle it by providing a custom exception.

Changes

Introduce InvalidStateException and use an expectation to map an invalid status to such exception