Better streaming input client generation.
Motivation:

The generated streaming input client code generates a method consuming a handler of GrpcReadStream of messages.

This is not correct nor sufficient, error is not signaled and this is not Rx/Mutiny friendly.

Changes:

Change the handler of GrpcReadStream for a completable, so the error is properly signalled.

Add an alternative method that consumes directly a ReadStream of messages, so this will be clearly identified by Rx/Mutiny generators that will have the opportunity to use a Flowable/Multi instead.