Improve transcoding integration.
Motivation:

Transcoding server requests are currently handled with a specific mount point acceptor for specific transcoding service methods.

We can make the mountpoint handle only the grpc server router mounting and leave the transcoding invocation in a separate SPI for creating transcoding invocation.

This will allow the transcoding invoker to handle transcoded requests without the need of specific metadata and simplify the deployment of transcoded handlers.

Changes:

The mount point now provides a list of paths to mount in the server internal router. When a transcoded request comes in, the invoker provides a transcoded using the service method metadata to handle the request/response messages.

Transcoded requests also now uses the new unified message encoder/decoder so they can decode protobuf as well as json.