Proper implementation of service error mapping to status mapping.
Motivation:

The current mapping of errors to status is currently not homogeneous (between vertx/io service generation) and lacks of a status exception allowing to set force a status to be set by the wrapper.

Changes:

Introduce io.vertx.grpc.server.StatusException that conveys the grpc status to be returned by the service when this exception is encountered.