In a grpc-io client stub pass the vertx context when building a derived client
Motivation:

When a grpc-io client stub builds a new context, it attempts to get the vertx context from the channel with a cast. This cast fails when the channel has been rewrapped by the abstract stub. Instead the context should simply be propagated by the current stub instance.

Changes:

Pass the current stub context when deriving a stub.