Struct JSON parser.
Motivation:

Parsing a protobuf encoded message of struct.proto can be quite a common thing and has unwanted side effects such as depending on the protobuf library and creating an intermediary struct data structure.

Since struct.proto is quite well defined, we can implement a parser for it based on a simple decoder that tokenizes a protobuf structure.