Provide a ground for encoding/decoding protobuf.
This assumes a minimalist representation of a protobuf schema.

A reader/parser interact with user code through a visitor interface. The visitor interface signal protobuf events.

Reading emits events to the user implemented visitor that can reconstruct an object graph.

Writing relies on the using the visitor to emit events and get the writer writes the protobuf bytes.

Writing is done in two phases, the first phase computes the various messages lengths, the second phase emits bytes.

A proof of concept is implemented with Vert.x Json elements and the Google Struct message that can encode/decode Vert.x Json elements to protobuf Struct format.

A Google Struct is also implemented to see what a generated data structure close to the proto definitions takes to implement as visitor.