GrpcIo server bridge should resume reading message when sending an error.
Motivation:

The GrpcIo server bridge does not resume reading message when sending an error.

When an error happens in StreamObserver.onNext, the caller does not request a new message leading to stopping the consumption of messages.

Although the grpc response is properly sent to the client, this can lead to a dead-lock when the client sends many messages.

Changes:

When the GrpcIo server bridge sends a response, resume the stream adapter to release messages.

Avoid dispatching to StreamObserver.onNext when the stream is closed.